backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
pydantic==1.10.13
flask-jwt-extended==4.6.0
psycopg2==2.9.5
python-magic==0.4.27
werkzeug==3.0.1
flask-login==0.6.3
flask-mail==0.9.1
flask-wtf==1.2.1
email-validator==2.1.0.post1 